# 🎯 Sprint 5: Integration & Validation Complete - Core Component Decoupling Initiative