… SPRINT5_COMPLETION_REPORT.md
… docs/Architecture.md
… docs/Project_Success_Metrics.md
… docs/sprint5_implementation_summary.md
… tests/test_performance_validation.cpp